+++ src/index.ts
@@ -71,7 +71,15 @@ export { point, Point, PointObject, PointValue } from './lib/Point';
 export { polyline, Polyline, PolylineOptions, PolylineValue } from './lib/Polyline';
 export { polylineCollection, PolylineCollection } from './lib/PolylineCollection';
 export { polylineIcon, PolylineIcon, PolylineIconOptions, PolylineIconValue } from './lib/PolylineIcon';
-export { closeAllPopups, popup, Popup, PopupOptions, PopupValue } from './lib/Popup';
+export {
+    closeAllPopups,
+    DataPopupContent,
+    DataPopupValue,
+    popup,
+    Popup,
+    PopupOptions,
+    PopupValue,
+} from './lib/Popup';
 export { size, Size, SizeObject, SizeValue } from './lib/Size';
 export { svgSymbol, SvgSymbol, SvgSymbolOptions, SvgSymbolValue } from './lib/SvgSymbol';
 export { tooltip, Tooltip, TooltipOptions, TooltipValue } from './lib/Tooltip';